backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
